Made a vue template for the robot card component in the Robots page
